64bit portability fixes.

+++ include/common/GContainers.h
@@ -357,8 +357,8 @@ public:
 	~GStringPipe() {}
 
 	virtual int Pop(char *Str, int Chars);
-	virtual int Push(const char *Str, int Chars = -1);
-	virtual int Push(const char16 *Str, int Chars = -1);
+	virtual int Push(const char *Str, ssize_t Chars = -1);
+	virtual int Push(const char16 *Str, ssize_t Chars = -1);
 	char *NewStr() { return (char*)New(sizeof(char)); }
 	GString NewGStr();
 	char16 *NewStrW() { return (char16*)New(sizeof(char16)); }
