+++ include/common/GLexCpp.h
@@ -5,12 +5,12 @@
 #include "GString.h"
 
 /// The callback type for string duplication while lexing
-typedef char16 *(*LexCppStrdup)(void *Context, const char16 *Str, int Chars);
+typedef char16 *(*LexCppStrdup)(void *Context, const char16 *Str, ssize_t Chars);
 
 /// Default callback if you want a heap string
-extern char16 *LexStrdup(void *context, const char16 *start, int len);
+extern char16 *LexStrdup(void *context, const char16 *start, ssize_t len);
 /// Default callback if you don't need the string
-extern char16 *LexNoReturn(void *context, const char16 *start, int len);
+extern char16 *LexNoReturn(void *context, const char16 *start, ssize_t len);
 
 /// Converts the next lexical component into a string token
 extern char16 *LexCpp
