+++ include/common/GTextView3.h
@@ -61,9 +61,9 @@ public:
 		/// will leave alone.
 		int Owner;
 		/// The start index into the text buffer of the region to style.
-		int Start;
+		ssize_t Start;
 		/// The length of the styled region
-		int Len;
+		ssize_t Len;
 		/// The font to draw the styled text in
 		GFont *Font;
 		/// The colour to draw with. If transparent, then the default 
@@ -102,7 +102,7 @@ public:
 		#endif
 		virtual CURSOR_CHAR GetCursor()  { return 0; }
 
-		int End() const { return Start + Len; }
+		size_t End() const { return Start + Len; }
 
 		/// \returns true if style is the same
 		bool operator ==(const GStyle &s)
@@ -121,7 +121,7 @@ public:
 		}
 
 		/// Returns true if this style overlaps the position of 's'
-		bool Overlap(int sStart, int sLen)
+		bool Overlap(ssize_t sStart, ssize_t sLen)
 		{
 			if (sStart + sLen - 1 < Start ||
 				sStart >= Start + Len)
@@ -160,8 +160,8 @@ protected:
 	class GTextLine
 	{
 	public:
-		int Start;		// Start offset
-		int Len;		// length of text
+		size_t Start;	// Start offset
+		size_t Len;		// length of text
 		GRect r;		// Screen location
 		GColour c;		// Colour of line... transparent = default colour
 		GColour Back;	// Background colour or transparent
@@ -194,7 +194,7 @@ protected:
 
 	GFont *FixedFont;
 	int LineY;
-	int SelStart, SelEnd;
+	ssize_t SelStart, SelEnd;
 	int DocOffset;
 	int MaxX;
 	bool Blink;
@@ -209,22 +209,22 @@ protected:
 
 	// Data
 	char16 *Text;
-	int Cursor;
-	int Size;
-	int Alloc;
+	size_t Cursor;
+	size_t Size;
+	size_t Alloc;
 
 	// Undo stuff
 	bool UndoOn;
 	GUndo UndoQue;
 
 	// private methods
-	GTextLine *GetTextLine(int Offset, int *Index = 0);
+	GTextLine *GetTextLine(ssize_t Offset, int *Index = 0);
 	int SeekLine(int Start, GTextViewSeek Where);
 	int TextWidth(GFont *f, char16 *s, int Len, int x, int Origin);
 	int ScrollYLine();
 	int ScrollYPixel();
 	GRect DocToScreen(GRect r);
-	int MatchText(char16 *Text, bool MatchWord, bool MatchCase, bool SelectionOnly);
+	ptrdiff_t MatchText(char16 *Text, bool MatchWord, bool MatchCase, bool SelectionOnly);
 	
 	// styles
 	bool InsertStyle(GAutoPtr<GStyle> s);
@@ -234,11 +234,11 @@ protected:
 	int SpaceDepth(char16 *Start, char16 *End);
 
 	// Overridables
-	virtual void PourText(int Start, int Length);
-	virtual void PourStyle(int Start, int Length);
+	virtual void PourText(size_t Start, ssize_t Length);
+	virtual void PourStyle(size_t Start, ssize_t Length);
 	virtual void OnFontChange();
 	virtual void OnPaintLeftMargin(GSurface *pDC, GRect &r, GColour &colour);
-	virtual char16 *MapText(char16 *Str, int Len, bool RtlTrailingSpace = false);
+	virtual char16 *MapText(char16 *Str, ssize_t Len, bool RtlTrailingSpace = false);
 
 	#ifdef _DEBUG
 	// debug
@@ -267,7 +267,7 @@ public:
 	int64 Value();
 	void Value(int64 i);
 	const char *GetMimeType() { return "text/plain"; }
-	int GetSize() { return Size; }
+	size_t GetSize() { return Size; }
 	GString operator[](int LineIdx);
 
 	int HitText(int x, int y, bool Nearest);
@@ -287,15 +287,15 @@ public:
 	void SetWrapType(uint8 i);
 	
 	// State / Selection
-	void SetCursor(int i, bool Select, bool ForceFullUpdate = false);
-	int IndexAt(int x, int y);
+	void SetCursor(size_t i, bool Select, bool ForceFullUpdate = false);
+	ssize_t IndexAt(int x, int y);
 	bool IsDirty() { return Dirty; }
 	void IsDirty(bool d) { Dirty = d; }
 	bool HasSelection();
 	void UnSelectAll();
 	void SelectWord(int From);
 	void SelectAll();
-	int GetCursor(bool Cursor = true);
+	ssize_t GetCursor(bool Cursor = true);
 	bool GetLineColumnAtIndex(GdcPt2 &Pt, int Index = -1);
 	int GetLines();
 	void GetTextExtent(int &x, int &y);
@@ -357,8 +357,8 @@ public:
 	LgiCursor GetCursor(int x, int y) { return LCUR_Ibeam; }
 
 	// Virtuals
-	virtual bool Insert(int At, char16 *Data, int Len);
-	virtual bool Delete(int At, int Len);
+	virtual bool Insert(size_t At, char16 *Data, ssize_t Len);
+	virtual bool Delete(size_t At, ssize_t Len);
 	virtual void OnEnter(GKey &k);
 	virtual void OnUrl(char *Url);
 	virtual void DoContextMenu(GMouse &m);
