+++ include/common/LgiCommon.h
@@ -63,7 +63,7 @@ extern "C"
 /// Converts a buffer of text to a different charset
 /// \ingroup Text
 /// \returns the bytes written to the location pointed to by 'Out'
-LgiFunc int LgiBufConvertCp(void *Out, const char *OutCp, int OutLen, const void *&In, const char *InCp, ptrdiff_t &InLen);
+LgiFunc ssize_t LgiBufConvertCp(void *Out, const char *OutCp, ssize_t OutLen, const void *&In, const char *InCp, ssize_t &InLen);
 
 /// \brief Converts a string to a new charset
 /// \return A dynamically allocate, null terminated string in the new charset
@@ -77,7 +77,7 @@ LgiFunc void *LgiNewConvertCp
 	/// The input data's charset
 	const char *InCp,
 	/// Bytes of valid data in the input
-	ptrdiff_t InLen = -1
+	ssize_t InLen = -1
 );
 
 /// Return true if Lgi support the charset
@@ -99,7 +99,7 @@ LgiFunc char *LgiSeekUtf8
 	/// Pointer to the current character
 	const char *Ptr,
 	/// The number of characters to move forward or back
-	int D,
+	ssize_t D,
 	/// The start of the memory buffer if you known
 	char *Start = 0
 );
